requirements.txt
pyyaml==6.0.2
pydantic==2.9.2
ujson==5.10.0
